requirements.txt
# requirements for ollama model manager package ollamacide

pytest
Pillow
